+++ Terminal/TerminalLib/Logic/Logic.h
@@ -227,7 +227,7 @@ class CLogic : CLogicAbstract, public ILogic
 
 	virtual void write_eeprom(byte cell_number, uint32_t value) final;
 
-	virtual void time_and_money(int16_t time, int16_t money) final;
+	virtual void time_and_money(int16_t time, int32_t money) final;
 
 	virtual void close_valve(byte number) final;
